Added flashlight and batteries
Added a flashlight which is not attached to player camera so that has an offset.

No flashlight model yet.

Flashlight has a lifetime. Turns off when reaches 0%.

Added text to display the lifetime and number of batteries player has.

Player can pickup batteries. Press R to reload. Press E to collect.