Added locked boxes and doors
Added locked boxes that require keys to open.

The boxes contain a key that can used to open locked door.

The right key is required to unlock the door.

Each key is a duplication in the hierarchy.
When box key is selected, it becomes active in the player's inventory, when active, it can be used to unlock the box. Same goes for the door