Added keypad door
This basement door required a keycode to be entered into the keypad.

Press E to view the keypad. Use left mouse button to click the numbers.

Once done, click execute to confirm. If wrong code is entered, a sound will be played. If correct, sound will be played and door will be opened.