Added Enemy AI
The enemy AI now patrols the basement maze.

It will walk to 4 destinations. At the destination, it will idle for a few seconds. This part is a bit buggy, it walks a few steps away from the destination and idles again.

It chases the player when it enters the look trigger zone. Right now the look zone is just a rectangular area infront of the AI.